[#133] refactor: notification service 구조 변경
