+++ src/main/java/org/silsagusi/joonggaemoa/api/notify/domain/entity/Notification.java
@@ -1,4 +1,4 @@
-package org.silsagusi.joonggaemoa.api.notify.domain;
+package org.silsagusi.joonggaemoa.api.notify.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
