+++ src/main/java/org/silsagusi/joonggaemoa/api/notify/domain/entity/NotificationType.java
@@ -1,4 +1,4 @@
-package org.silsagusi.joonggaemoa.api.notify.domain;
+package org.silsagusi.joonggaemoa.api.notify.domain.entity;
 
 public enum NotificationType {
     CONNECTION,
