+++ src/main/java/org/silsagusi/joonggaemoa/api/notify/infrastructure/NotificationRepository.java
@@ -1,7 +1,7 @@
 package org.silsagusi.joonggaemoa.api.notify.infrastructure;
 
-import org.silsagusi.joonggaemoa.api.notify.domain.Notification;
-import org.silsagusi.joonggaemoa.api.notify.domain.NotificationType;
+import org.silsagusi.joonggaemoa.api.notify.domain.entity.Notification;
+import org.silsagusi.joonggaemoa.api.notify.domain.entity.NotificationType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -11,4 +11,5 @@ public interface NotificationRepository extends JpaRepository<Notification, Long
 
     boolean existsByAgentIdAndTypeAndContent(Long agentId, NotificationType type, String content);
 
+    List<Notification> findAllByAgentId(Long agentId);
 }
