+++ src/main/java/org/silsagusi/joonggaemoa/global/batch/ContractExpireNotificationJobConfig.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.silsagusi.joonggaemoa.api.contract.domain.entity.Contract;
 import org.silsagusi.joonggaemoa.api.contract.infrastructure.ContractRepository;
 import org.silsagusi.joonggaemoa.api.notify.application.NotificationService;
-import org.silsagusi.joonggaemoa.api.notify.domain.NotificationType;
+import org.silsagusi.joonggaemoa.api.notify.domain.entity.NotificationType;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobScope;
